move package to root folder
