+++ init/resources/hasura/endpoints/cicd_artifact.gql
@@ -3,7 +3,7 @@ mutation insert_cicd_Artifact(
   $data_artifact_id: String!
   $data_artifact_organization: String!
   $data_artifact_source: String!
-  $data_origin: String!
+  $data_origin: String
 ) {
   insert_cicd_Artifact_one(
     object: {
