+++ init/resources/hasura/endpoints/cicd_artifact_commit_association.gql
@@ -7,7 +7,7 @@ mutation insert_cicd_artifact_commit_association(
   $data_commit_sha: String!
   $data_commit_source: String!
   $data_artifact_source: String!
-  $data_origin: String!
+  $data_origin: String
 ) {
   insert_cicd_ArtifactCommitAssociation_one(
     object: {
