+++ init/resources/hasura/endpoints/cicd_artifact_with_build.gql
@@ -7,7 +7,7 @@ mutation insert_cicd_Artifact_WithBuildKey(
   $data_artifact_organization: String!
   $data_run_id: String!
   $data_artifact_source: String!
-  $data_origin: String!
+  $data_origin: String
 ) {
   insert_cicd_Artifact_one(
     object: {
