+++ init/resources/hasura/endpoints/cicd_deployment.gql
@@ -6,7 +6,7 @@ mutation insert_cicd_deployment_with_start_end(
   $env: jsonb!
   $deploy_start_time: timestamptz!
   $deploy_end_time: timestamptz!
-  $data_origin: String!
+  $data_origin: String
 ) {
   insert_cicd_Deployment_one(
     object: {
