+++ init/resources/hasura/endpoints/cicd_organization.gql
@@ -1,7 +1,7 @@
 mutation insert_cicd_Organization(
   $data_artifact_organization: String!
   $data_artifact_source: String
-  $data_origin: String!
+  $data_origin: String
 ) {
   insert_cicd_Organization_one(
     object: {
