+++ init/resources/hasura/endpoints/cicd_organization_from_run.gql
@@ -1,7 +1,7 @@
 mutation insert_cicd_Organization(
   $data_run_organization: String!
   $data_run_source: String
-  $data_origin: String!
+  $data_origin: String
 ) {
   insert_cicd_Organization_one(
     object: {
