+++ init/resources/hasura/endpoints/cicd_repository.gql
@@ -2,7 +2,7 @@ mutation insert_cicd_Repository(
   $data_artifact_repository: String!
   $data_artifact_organization: String!
   $data_artifact_source: String!
-  $data_origin: String!
+  $data_origin: String
 ) {
   insert_cicd_Repository_one(
     object: {
