+++ init/resources/hasura/endpoints/delete_cicd_Organization.gql
@@ -1,4 +1,4 @@
-mutation delete_cicd_Organization($data_origin: String!) {
+mutation delete_cicd_Organization($data_origin: String) {
   delete_cicd_Organization(where: { origin: { _eq: $data_origin } }) {
     affected_rows
   }
