+++ init/resources/hasura/endpoints/delete_cicd_artifact_commit_association.gql
@@ -1,4 +1,4 @@
-mutation insert_vcs_pull_request_review($data_origin: String!) {
+mutation insert_vcs_pull_request_review($data_origin: String) {
   delete_cicd_ArtifactCommitAssociation(
     where: { origin: { _eq: $data_origin } }
   ) {
