+++ init/resources/hasura/endpoints/delete_cicd_artifact_deployment.gql
@@ -1,4 +1,4 @@
-mutation delete_cicd_artifact_deployment($data_origin: String!) {
+mutation delete_cicd_artifact_deployment($data_origin: String) {
   delete_cicd_ArtifactDeployment(where: { origin: { _eq: $data_origin } }) {
     affected_rows
   }
