+++ init/resources/hasura/endpoints/delete_compute_application.gql
@@ -1,4 +1,4 @@
-mutation delete_compute_application($data_origin: String!) {
+mutation delete_compute_application($data_origin: String) {
   delete_compute_Application(where: { origin: { _eq: $data_origin } }) {
     affected_rows
   }
