+++ init/resources/hasura/endpoints/delete_ims_incident_application_impact.gql
@@ -1,4 +1,4 @@
-mutation delete_ims_incident_application_impact($data_origin: String!) {
+mutation delete_ims_incident_application_impact($data_origin: String) {
   delete_ims_IncidentApplicationImpact(
     where: { origin: { _eq: $data_origin } }
   ) {
