+++ init/resources/hasura/endpoints/delete_vcs_commit.gql
@@ -1,4 +1,4 @@
-mutation delete_vcs_commit($data_origin: String!) {
+mutation delete_vcs_commit($data_origin: String) {
   delete_vcs_Commit(where: { origin: { _eq: $data_origin } }) {
     affected_rows
   }
