+++ init/resources/hasura/endpoints/delete_vcs_pull_request.gql
@@ -1,4 +1,4 @@
-mutation delete_vcs_pull_request($data_origin: String!) {
+mutation delete_vcs_pull_request($data_origin: String) {
   delete_vcs_PullRequest(where: { origin: { _eq: $data_origin } }) {
     affected_rows
   }
