+++ init/resources/hasura/endpoints/delete_vcs_pull_request_review.gql
@@ -1,4 +1,4 @@
-mutation delete_vcs_pull_request_review($data_origin: String!) {
+mutation delete_vcs_pull_request_review($data_origin: String) {
   delete_vcs_PullRequestReview(where: { origin: { _eq: $data_origin } }) {
     affected_rows
   }
