+++ init/resources/hasura/endpoints/delete_vcs_repository.gql
@@ -1,4 +1,4 @@
-mutation delete_vcs_repository($data_origin: String!) {
+mutation delete_vcs_repository($data_origin: String) {
   delete_vcs_Repository(where: { origin: { _eq: $data_origin } }) {
     affected_rows
   }
