+++ init/resources/hasura/endpoints/delete_vcs_user.gql
@@ -1,4 +1,4 @@
-mutation delete_vcs_user($data_origin: String!) {
+mutation delete_vcs_user($data_origin: String) {
 	delete_vcs_User(where: { origin: { _eq: $data_origin } }) {
 		affected_rows
 	}
