+++ mock-data/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "faros-ce-app-init",
+  "name": "faros-ce-mock-data",
   "version": "0.0.1",
-  "description": "Setup logic for Faros CE services",
+  "description": "Faros CE Mock Data Script",
   "author": "Faros AI, Inc.",
   "homepage": "https://www.faros.ai",
   "license": "UNLICENSED",
@@ -9,50 +9,29 @@
   "scripts": {
     "build": "tsc -p src",
     "clean": "rm -rf lib node_modules out",
-    "fix": "prettier --write 'src/**/*.ts' 'test/**/*.ts' && npm run lint -- --fix",
-    "lint": "eslint 'src/**/*.ts' 'test/**/*.ts'",
+    "fix": "prettier --write 'src/*.ts' && npm run lint -- --fix",
+    "lint": "eslint 'src/**/*.ts'",
     "prepare": "npm run build",
     "test": "jest --verbose --color",
     "test-cov": "jest --coverage --verbose --color",
     "watch": "tsc -b -w src test"
   },
-  "files": [
-    "lib/",
-    "resources/"
-  ],
+  "files": ["lib/"],
+  "main": "./lib",
   "dependencies": {
-    "analytics-node": "^6.0.0",
-    "async-retry": "^1.3.3",
     "axios": "^0.25.0",
-    "figlet": "^1.5.2",
-    "fs-extra": "^10.0.0",
-    "handlebars": "^4.7.7",
-    "lodash": "^4.17.21",
+    "luxon": "^2.3.1",
     "pino": "^7.6.4",
-    "pluralize": "^8.0.0",
-    "tar": "^6.1.11",
-    "uuid": "^8.3.2",
     "verror": "^1.10.0"
   },
   "devDependencies": {
-    "@types/analytics-node": "^3.1.7",
-    "@types/async-retry": "^1.4.3",
-    "@types/figlet": "^1.5.4",
-    "@types/fs-extra": "^9.0.13",
-    "@types/jest": "^27.4.1",
-    "@types/lodash": "^4.14.168",
     "@types/luxon": "^2.0.9",
     "@types/node": "^17.0.21",
-    "@types/pluralize": "^0.0.29",
-    "@types/tar": "^6.1.1",
-    "@types/uuid": "^8.3.4",
     "@types/verror": "^1.10.4",
     "@typescript-eslint/eslint-plugin": "^5.10.1",
     "@typescript-eslint/parser": "^5.10.1",
     "eslint": "^8.7.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
-    "jest": "^27.5.1",
-    "nock": "^13.2.4",
     "pino-pretty": "^7.5.1",
     "prettier": "2.5.1",
     "ts-jest": "^27.1.3",
