+++ mock-data/src/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../.tsconfig.json",
+  "compilerOptions": {
+    "composite": true
+  }
+}
